Rewrote metadata structure
Did a bunch of things, but because the compiler is throwing an error when I try to define enum bencode case .dict [bencode: bencode] as circular, I've had to rewrite it as .dict [String: bencode] in the intermediate representation.

To accomodate this, I had to rewrite metadata.swift (torrent.swift of old) and my decode function.

To ensure correctness, I also wrote a info dict hash function that extracts everything out of the info dict, because i realised i cant construct it from expected fields. So it comes directly out of the adding phase.

Added a bunch of tests for the rewritten structures, as well as the info dict/hash from the default.

Should add the test .torrent but will figure that out later